refactor ibgp and ebgp fabric module names to be clear it's vxlan
