+++ tests/integration/targets/nd_manage_fabric/tasks/fabric_ebgp.yaml
@@ -18,7 +18,7 @@
 # CLEANUP - Ensure clean state before tests
 #############################################################################
 - name: Clean up any existing test fabrics before starting tests
-  cisco.nd.nd_manage_fabric_ebgp:
+  cisco.nd.nd_manage_fabric_ebgp_vxlan:
     <<: *nd_info
     state: deleted
     config:
@@ -31,7 +31,7 @@
 # TEST 1: STATE MERGED - Create fabric using merged state
 #############################################################################
 - name: "TEST 1a: Create eBGP fabric using state merged (first run)"
-  cisco.nd.nd_manage_fabric_ebgp:
+  cisco.nd.nd_manage_fabric_ebgp_vxlan:
     <<: *nd_info
     state: merged
     config:
@@ -49,7 +49,7 @@
   tags: [test_merged, test_merged_create]
 
 - name: "TEST 1b: Create eBGP fabric using state merged (second run - idempotency test)"
-  cisco.nd.nd_manage_fabric_ebgp:
+  cisco.nd.nd_manage_fabric_ebgp_vxlan:
     <<: *nd_info
     state: merged
     config:
@@ -67,7 +67,7 @@
   tags: [test_merged, test_merged_idempotent]
 
 - name: "TEST 1c: Update eBGP fabric using state merged (modify existing)"
-  cisco.nd.nd_manage_fabric_ebgp:
+  cisco.nd.nd_manage_fabric_ebgp_vxlan:
     <<: *nd_info
     state: merged
     config:
@@ -265,7 +265,7 @@
 # TEST 2: STATE REPLACED - Create and manage fabric using replaced state
 #############################################################################
 - name: "TEST 2a: Create eBGP fabric using state replaced (first run)"
-  cisco.nd.nd_manage_fabric_ebgp:
+  cisco.nd.nd_manage_fabric_ebgp_vxlan:
     <<: *nd_info
     state: replaced
     config:
@@ -366,7 +366,7 @@
   tags: [test_replaced, test_replaced_create]
 
 - name: "TEST 2b: Create eBGP fabric using state replaced (second run - idempotency test)"
-  cisco.nd.nd_manage_fabric_ebgp:
+  cisco.nd.nd_manage_fabric_ebgp_vxlan:
     <<: *nd_info
     state: replaced
     config:
@@ -467,7 +467,7 @@
   tags: [test_replaced, test_replaced_idempotent]
 
 - name: "TEST 2c: Update eBGP fabric using state replaced (complete replacement with minimal config)"
-  cisco.nd.nd_manage_fabric_ebgp:
+  cisco.nd.nd_manage_fabric_ebgp_vxlan:
     <<: *nd_info
     state: replaced
     config:
@@ -897,7 +897,7 @@
 # TEST 3: Demonstrate difference between merged and replaced states
 #############################################################################
 - name: "TEST 3: Create eBGP fabric for merged vs replaced comparison"
-  cisco.nd.nd_manage_fabric_ebgp:
+  cisco.nd.nd_manage_fabric_ebgp_vxlan:
     <<: *nd_info
     state: replaced
     config:
@@ -906,7 +906,7 @@
   tags: [test_comparison]
 
 - name: "TEST 3a: Partial update using merged state (should merge changes)"
-  cisco.nd.nd_manage_fabric_ebgp:
+  cisco.nd.nd_manage_fabric_ebgp_vxlan:
     <<: *nd_info
     state: merged
     config:
@@ -930,7 +930,7 @@
   tags: [test_comparison, test_merged_partial]
 
 - name: "TEST 3b: Partial update using replaced state (should replace entire config)"
-  cisco.nd.nd_manage_fabric_ebgp:
+  cisco.nd.nd_manage_fabric_ebgp_vxlan:
     <<: *nd_info
     state: replaced
     config:
@@ -958,7 +958,7 @@
 # TEST 4: STATE DELETED - Delete fabrics
 #############################################################################
 - name: "TEST 4a: Delete eBGP fabric using state deleted"
-  cisco.nd.nd_manage_fabric_ebgp:
+  cisco.nd.nd_manage_fabric_ebgp_vxlan:
     <<: *nd_info
     state: deleted
     config:
@@ -976,7 +976,7 @@
   tags: [test_deleted, test_deleted_delete]
 
 - name: "TEST 4b: Delete eBGP fabric using state deleted (second run - idempotency test)"
-  cisco.nd.nd_manage_fabric_ebgp:
+  cisco.nd.nd_manage_fabric_ebgp_vxlan:
     <<: *nd_info
     state: deleted
     config:
@@ -997,7 +997,7 @@
 # TEST 5: Multiple fabric operations in single task
 #############################################################################
 - name: "TEST 5: Multiple eBGP fabric operations in single task"
-  cisco.nd.nd_manage_fabric_ebgp:
+  cisco.nd.nd_manage_fabric_ebgp_vxlan:
     <<: *nd_info
     state: merged
     config:
@@ -1175,7 +1175,7 @@
 # FINAL CLEANUP - Clean up all test fabrics
 #############################################################################
 - name: "CLEANUP: Delete all test eBGP fabrics"
-  cisco.nd.nd_manage_fabric_ebgp:
+  cisco.nd.nd_manage_fabric_ebgp_vxlan:
     <<: *nd_info
     state: deleted
     config:
@@ -1194,7 +1194,7 @@
   debug:
     msg: |
       ========================================================
-      TEST SUMMARY for cisco.nd.nd_manage_fabric_ebgp module:
+      TEST SUMMARY for cisco.nd.nd_manage_fabric_ebgp_vxlan module:
       ========================================================
       ✓ TEST 1: STATE MERGED
         - Create fabric: {{ 'PASSED' if ebgp_merged_result_1 is changed else 'FAILED' }}
